requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pinecone-client==5.0.1
openai==1.55.0
crewai==0.86.0
crewai-tools==0.14.0
pymupdf==1.24.0
python-docx==1.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
pydantic==2.10.0
python-multipart==0.0.12
requests==2.32.0